Update CInterface files
