+++ Engine/source/cinterface/cinterface.h
@@ -27,38 +27,7 @@
 
 // cinterface can override this (useful for plugins, etc)
 extern "C" {
-
-   const char* torque_getexecutablepath();
-   DLL_DECL void torque_reset();
-   DLL_DECL bool torque_engineinit(S32 argc, const char **argv);
-   DLL_DECL S32 torque_enginetick();
-   DLL_DECL S32 torque_engineshutdown();
-
-   DLL_DECL S32 torque_getreturnstatus();
-   DLL_DECL void torque_enginesignalshutdown();
-   DLL_DECL bool torque_isdebugbuild();
-
    DLL_DECL void SetCallbacks(void* ptr, void* methodPtr, void* isMethodPtr, void *mainPtr);
-   DLL_DECL SimObject* FindDataBlockByName(const char* pName);
-   DLL_DECL SimObject* FindObjectByName(const char* pName);
-   DLL_DECL SimObject* FindObjectById(U32 pId);
-   DLL_DECL SimObjectPtr<SimObject>* WrapObject(SimObject* pObject);
-   DLL_DECL void Sim_DeleteObjectPtr(SimObjectPtr<SimObject>* pObjectPtr);
-   DLL_DECL bool fnSimObject_registerObject(SimObject* pObject);
-
-   DLL_DECL const char* fn_getConsoleString(const char* name);
-   DLL_DECL void fn_setConsoleString(const char* name, const char* value);
-   DLL_DECL S32 fn_getConsoleInt(const char* name);
-   DLL_DECL void fn_setConsoleInt(const char* name, S32 value);
-   DLL_DECL F32 fn_getConsoleFloat(const char* name);
-   DLL_DECL void fn_setConsoleFloat(const char* name, F32 value);
-   DLL_DECL bool fn_getConsoleBool(const char* name);
-   DLL_DECL void fn_setConsoleBool(const char* name, bool value);
-   DLL_DECL void fnSimDataBlock_AssignId(SimDataBlock* db);
-   DLL_DECL void fnSimDataBlock_Preload(SimDataBlock* db);
-   DLL_DECL void fnSimObject_CopyFrom(SimObject* obj, SimObject* parent);
-   DLL_DECL void fnSimObject_SetMods(SimObject* obj, bool modStaticFields, bool modDynamicFields);
-
 }
 
 #define CALL_CINTERFACE_FUNCTION(name, ...){const char *v[] = { __VA_ARGS__ }; CInterface::CallFunction(name, v, sizeof(v) / sizeof(*v));}
@@ -72,10 +41,10 @@ class CInterface {
    CallFunctionCallback mFunctionCallback;
    CallMethodCallback mMethodCallback;
    CallMainCallback mMainCallback;
-   const char* _CallFunction(const char* nameSpace, const char* name, const char **argv, int argc, bool *result);
-   const char* _CallMethod(const char* className, const char* classNamespace, U32 object, const char* name, const char **argv, int argc, bool *res);
-   void _CallMain(bool *res);
-   bool _isMethod(const char* className, const char* methodName);
+   const char* _CallFunction(const char* nameSpace, const char* name, const char **argv, int argc, bool *result) const;
+   const char* _CallMethod(const char* className, const char* classNamespace, U32 object, const char* name, const char **argv, int argc, bool *res) const;
+   void _CallMain(bool *res) const;
+   bool _isMethod(const char* className, const char* methodName) const;
 public:
    CInterface()
    {
